Refactor emoji message handling & panel lifecycle
Move chat/message-related emoji logic out of the main plugin into a dedicated EmojiMessageManager service and add a panel manager to handle the plugin panel lifecycle. Introduce GitHubDownloadStarted and GitHubDownloadCompleted events and update GitHubEmojiDownloader to emit those events, react to config changes, and report results to the client/UI. Update CustomEmojiPanel and EmojiTreePanel to react to download events and use the EventBus; add EmojiPanelManager to create/remove the navigation button and manage the panel. Add config change handling in EmojiLoader (max image height) and wire the new services into LifecycleManager so lifecycle management is centralized. Overall this decouples responsibilities, improves lifecycle control, and makes GitHub download/reporting flow event-driven.
