+++ src/main/java/com/customemoji/event/GitHubDownloadCompleted.java
@@ -1,0 +1,11 @@
+package com.customemoji.event;
+
+import com.customemoji.io.GitHubEmojiDownloader;
+import lombok.Value;
+
+@Value
+public class GitHubDownloadCompleted
+{
+	GitHubEmojiDownloader.DownloadResult result;
+	boolean hadPreviousDownload;
+}
