+++ src/main/java/com/customemoji/event/GitHubDownloadStarted.java
@@ -1,0 +1,8 @@
+package com.customemoji.event;
+
+import lombok.Value;
+
+@Value
+public class GitHubDownloadStarted
+{
+}
